split data loader test & coverage ci fix, test=develop
