+++ python/paddle/fluid/tests/unittests/test_imperative_data_loader.py
@@ -12,15 +12,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import os
 import sys
-import signal
 import unittest
 import numpy as np
-
 import paddle.fluid as fluid
-from paddle.fluid import core
-import paddle.compat as cpt
 
 if sys.version_info[0] == 2:
     import Queue as queue
@@ -72,13 +67,7 @@ class TestDygraphhDataLoader(unittest.TestCase):
         self.batch_num = 4
         self.epoch_num = 2
 
-    def test_not_capacity(self):
-        with fluid.dygraph.guard():
-            with self.assertRaisesRegexp(ValueError,
-                                         "Please give value to capacity."):
-                fluid.io.DataLoader.from_generator()
-
-    def test_single_process(self):
+    def test_single_process_reader(self):
         with fluid.dygraph.guard():
             loader = fluid.io.DataLoader.from_generator(
                 capacity=2, iterable=False, use_multiprocess=False)
@@ -146,52 +135,6 @@ class TestDygraphhDataLoader(unittest.TestCase):
             loader._reader_process_loop()
             for _ in range(self.batch_num):
                 loader._data_queue.get(timeout=10)
-
-    # NOTE: exception tests
-    def test_single_process_with_thread_expection(self):
-        def error_sample_genarator(batch_num):
-            def __reader__():
-                for _ in range(batch_num):
-                    yield [[[1, 2], [1]]]
-
-            return __reader__
-
-        with fluid.dygraph.guard():
-            loader = fluid.io.DataLoader.from_generator(
-                capacity=2, iterable=False, use_multiprocess=False)
-            loader.set_batch_generator(
-                error_sample_genarator(self.batch_num), places=fluid.CPUPlace())
-            exception = None
-            try:
-                for _ in loader():
-                    print("test_single_process_with_thread_expection")
-            except core.EnforceNotMet as ex:
-                self.assertIn("Blocking queue is killed",
-                              cpt.get_exception_message(ex))
-                exception = ex
-            self.assertIsNotNone(exception)
-
-    def test_multi_process_with_thread_expection(self):
-        def error_sample_genarator(batch_num):
-            def __reader__():
-                for _ in range(batch_num):
-                    yield [[[1, 2], [1]]]
-
-            return __reader__
-
-        with fluid.dygraph.guard():
-            loader = fluid.io.DataLoader.from_generator(capacity=2)
-            loader.set_batch_generator(
-                error_sample_genarator(self.batch_num), places=fluid.CPUPlace())
-            exception = None
-            try:
-                for _ in loader():
-                    print("test_multi_process_with_thread_expection")
-            except core.EnforceNotMet as ex:
-                self.assertIn("Blocking queue is killed",
-                              cpt.get_exception_message(ex))
-                exception = ex
-            self.assertIsNotNone(exception)
 
 
 if __name__ == '__main__':
